+++ vortex-array/src/arrays/decimal/compute/sum.rs
@@ -1,11 +1,10 @@
-use itertools::Itertools;
 use vortex_error::{VortexResult, vortex_bail};
 use vortex_mask::Mask;
-use vortex_scalar::Scalar;
+use vortex_scalar::{Scalar, match_each_decimal_value_type};
 
 use crate::arrays::{DecimalArray, DecimalEncoding};
 use crate::compute::{SumKernel, SumKernelAdapter};
-use crate::{Array, match_each_decimal_value_type, register_kernel};
+use crate::{Array, register_kernel};
 
 macro_rules! sum_decimal {
     ($ty:ty, $values:expr) => {{
@@ -16,6 +15,8 @@ macro_rules! sum_decimal {
         sum
     }};
     ($ty:ty, $values:expr, $validity:expr) => {{
+        use itertools::Itertools;
+
         let mut sum: $ty = <$ty>::default();
         for (v, valid) in $values.iter().zip_eq($validity.iter()) {
             if valid {
