+++ vortex-array/src/builders/decimal.rs
@@ -4,8 +4,9 @@ use vortex_buffer::BufferMut;
 use vortex_dtype::{DType, DecimalDType, Nullability};
 use vortex_error::{VortexResult, vortex_bail, vortex_panic};
 use vortex_mask::Mask;
+use vortex_scalar::NativeDecimalType;
 
-use crate::arrays::{BoolArray, DecimalArray, NativeDecimalType};
+use crate::arrays::{BoolArray, DecimalArray};
 use crate::builders::ArrayBuilder;
 use crate::builders::lazy_validity_builder::LazyNullBufferBuilder;
 use crate::validity::Validity;
