+++ vortex-array/src/builders/mod.rs
@@ -49,12 +49,12 @@ use vortex_error::{VortexResult, vortex_bail, vortex_err};
 use vortex_mask::Mask;
 use vortex_scalar::{
     BinaryScalar, BoolScalar, ExtScalar, ListScalar, PrimitiveScalar, Scalar, ScalarValue,
-    StructScalar, Utf8Scalar,
+    StructScalar, Utf8Scalar, match_each_decimal_value_type,
 };
 
 use crate::arrays::precision_to_storage_size;
 use crate::builders::struct_::StructBuilder;
-use crate::{Array, ArrayRef, match_each_decimal_value_type};
+use crate::{Array, ArrayRef};
 
 pub trait ArrayBuilder: Send {
     fn as_any(&self) -> &dyn Any;
