+++ vortex-scalar/src/decimal.rs
@@ -8,6 +8,19 @@ use vortex_error::{VortexError, VortexResult, vortex_bail};
 use crate::scalar_value::InnerScalarValue;
 use crate::{Scalar, ScalarValue, i256};
 
+/// Type of the decimal values.
+#[derive(Clone, Copy, Debug, prost::Enumeration, PartialEq, Eq)]
+#[repr(u8)]
+#[non_exhaustive]
+pub enum DecimalValueType {
+    I8 = 0,
+    I16 = 1,
+    I32 = 2,
+    I64 = 3,
+    I128 = 4,
+    I256 = 5,
+}
+
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd)]
 pub enum DecimalValue {
     I8(i8),
@@ -16,6 +29,35 @@ pub enum DecimalValue {
     I64(i64),
     I128(i128),
     I256(i256),
+}
+
+/// Type of decimal scalar values.
+pub trait NativeDecimalType: Copy + Eq + Ord {
+    const VALUES_TYPE: DecimalValueType;
+}
+
+impl NativeDecimalType for i8 {
+    const VALUES_TYPE: DecimalValueType = DecimalValueType::I8;
+}
+
+impl NativeDecimalType for i16 {
+    const VALUES_TYPE: DecimalValueType = DecimalValueType::I16;
+}
+
+impl NativeDecimalType for i32 {
+    const VALUES_TYPE: DecimalValueType = DecimalValueType::I32;
+}
+
+impl NativeDecimalType for i64 {
+    const VALUES_TYPE: DecimalValueType = DecimalValueType::I64;
+}
+
+impl NativeDecimalType for i128 {
+    const VALUES_TYPE: DecimalValueType = DecimalValueType::I128;
+}
+
+impl NativeDecimalType for i256 {
+    const VALUES_TYPE: DecimalValueType = DecimalValueType::I256;
 }
 
 impl Display for DecimalValue {
@@ -212,3 +254,53 @@ decimal_scalar_pack!(u64, i128, I128);
 
 decimal_scalar_pack!(i128, i128, I128);
 decimal_scalar_pack!(i256, i256, I256);
+
+#[macro_export]
+macro_rules! match_each_decimal_value {
+    ($self:expr, | $_:tt $value:ident | $($body:tt)*) => ({
+        macro_rules! __with__ {( $_ $value:ident ) => ( $($body)* )}
+        macro_rules! __with__ {( $_ $value:ident ) => ( $($body)* )}
+        match $self {
+            DecimalValue::I8(v) => __with__! { v },
+            DecimalValue::I16(v) => __with__! { v },
+            DecimalValue::I32(v) => __with__! { v },
+            DecimalValue::I64(v) => __with__! { v },
+            DecimalValue::I128(v) => __with__! { v },
+            DecimalValue::I256(v) => __with__! { v },
+        }
+    });
+}
+
+/// Macro to match over each decimal value type, binding the corresponding native type (from `DecimalValueType`)
+#[macro_export]
+macro_rules! match_each_decimal_value_type {
+    ($self:expr, | $_:tt $enc:ident | $($body:tt)*) => ({
+        macro_rules! __with__ {( $_ $enc:ident ) => ( $($body)* )}
+        use $crate::{DecimalValueType, i256};
+        match $self {
+            DecimalValueType::I8 => __with__! { i8 },
+            DecimalValueType::I16 => __with__! { i16 },
+            DecimalValueType::I32 => __with__! { i32 },
+            DecimalValueType::I64 => __with__! { i64 },
+            DecimalValueType::I128 => __with__! { i128 },
+            DecimalValueType::I256 => __with__! { i256 },
+            ty => vortex_error::vortex_panic!("unknown decimal value type {:?}", ty),
+        }
+    });
+    ($self:expr, | ($_0:tt $enc:ident, $_1:tt $dv_path:ident) | $($body:tt)*) => ({
+        macro_rules! __with2__ { ( $_0 $enc:ident, $_1 $dv_path:ident ) => ( $($body)* ) }
+        use $crate::{DecimalValueType, i256};
+        use $crate::DecimalValue::*;
+
+        match $self {
+            DecimalValueType::I8 => __with2__! { i8, I8 },
+            DecimalValueType::I16 => __with2__! { i16, I16 },
+            DecimalValueType::I32 => __with2__! { i32, I32 },
+            DecimalValueType::I64 => __with2__! { i64, I64 },
+            DecimalValueType::I128 => __with2__! { i128, I128 },
+            DecimalValueType::I256 => __with2__! { i256, I256 },
+            ty => vortex_error::vortex_panic!("unknown decimal value type {:?}", ty),
+
+        }
+    });
+}
